feat: 결제 창 까지 로드 되도록 구현

+++ src/main/java/com/bootcamp/paymentdemo/domain/order/repository/OrderRepository.java
@@ -3,7 +3,5 @@ package com.bootcamp.paymentdemo.domain.order.repository;
 import com.bootcamp.paymentdemo.domain.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-
-
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
